Refactor: store room in walkable map; clean up types
Instead of linearly searching over all rooms to find a room with a
given position, it's easier to put the room into the walkable lookup
map. I also put doors in there.

The walkable map is used during pathfinding. To calculate movement
costs, I'll need to know the room and its furniture, so this change to
the walkable map will help implement pathfinding weights.

During the refactor I noticed that some of the type errors weren't
being caught. The problem was that I had put types into mapgen.d.ts
but that *prevented* typescript from using the mapgen.js's types when
importing that module. I moved the types to types.d.ts and it started
finding more types in the js code. I fixed some type declarations and
added more.