Feat: production jobs, and debug output tables
For testing I wanted three types of rooms:

1. Farm: no inputs, with outputs
2. Dining: with inputs, no outputs
3. Kitchen: with inputs, with outputs

This was tricky to get working. There's an implicit state machine in
how the colonists work, and that was error prone. There were also several
bugs involving whether an item or colonist can be assigned to a job.

I ended up assigning IDs to each item and also displayed items and
colonists in a debug table so that I could figure out what was going
on.